+++ src/MicroElements.Metadata/Metadata/Formatters/DefaultToStringFormatter.cs
@@ -1,8 +1,15 @@
 ﻿// Copyright (c) MicroElements. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using System;
+using System.Globalization;
+
 namespace MicroElements.Metadata.Formatters
 {
+    /// <summary>
+    /// Default ToString formatter.
+    /// Uses <see cref="object.ToString"/> method.
+    /// </summary>
     public class DefaultToStringFormatter : IValueFormatter<object>
     {
         /// <summary>
@@ -11,6 +18,16 @@ namespace MicroElements.Metadata.Formatters
         public static IValueFormatter<object> Instance { get; } = new DefaultToStringFormatter();
 
         /// <inheritdoc />
-        public string? Format(object? value) => value?.ToString();
+        public string? Format(object? value)
+        {
+            if (value is null)
+                return null;
+
+            /* NOTE: Nullable types boxed as its underlying type. So no need to process nullability case. */
+
+            return value is IFormattable formattable
+                ? formattable.ToString(null, CultureInfo.InvariantCulture)
+                : value.ToString();
+        }
     }
 }