+++ src/MicroElements.Metadata/Metadata/Formatters/Formatter.cs
@@ -3,15 +3,28 @@
 
 namespace MicroElements.Metadata.Formatters
 {
-    public static partial class Formatter
+    public static class Formatter
     {
-        public static IValueFormatter ValueFormatter = new CompositeFormatter(new DefaultFormatProvider());
+        /// <summary>
+        /// Gets default single item formatter.
+        /// </summary>
+        public static IValueFormatter SingleValueFormatter { get; } = new CompositeFormatter(new DefaultFormatProvider().GetFormatters());
 
-        public static IValueFormatter DefaultToStringFormatter = new CompositeFormatter(new CollectionFormatter(ValueFormatter), ValueFormatter);
-    }
+        /// <summary>
+        /// Gets FullToStringFormatter that formats with <see cref="SingleValueFormatter"/> and can format collections.
+        /// </summary>
+        public static IValueFormatter FullToStringFormatter { get; } = new CompositeFormatter()
+            .With(SingleValueFormatter)
+            .With(new CollectionFormatter(SingleValueFormatter))
+            .With(new KeyValuePairFormatter(SingleValueFormatter))
+            .With(new ValueTuplePairFormatter(SingleValueFormatter))
+            .With(DefaultToStringFormatter.Instance);
 
-    public static partial class Formatter
-    {
-
+        public static IValueFormatter FullRecursiveFormatter { get; } = new CompositeFormatter()
+            .With(new DefaultFormatProvider().GetFormatters())
+            .With(new CollectionFormatter(new RuntimeFormatter(() => FullRecursiveFormatter)))
+            .With(new KeyValuePairFormatter(new RuntimeFormatter(() => FullRecursiveFormatter)))
+            .With(new ValueTuplePairFormatter(new RuntimeFormatter(() => FullRecursiveFormatter)))
+            .With(DefaultToStringFormatter.Instance);
     }
 }