+++ src/MicroElements.Metadata/Metadata/Formatters/NodaTimeIsoFormatters.cs
@@ -5,22 +5,35 @@ using System.Collections.Generic;
 
 namespace MicroElements.Metadata.Formatters
 {
+    /// <summary>
+    /// NodaTime formatters.
+    /// </summary>
     public class NodaTimeIsoFormatters : IValueFormatterProvider
     {
-        public static IValueFormatter LocalDateFormatter =
-            new FormattableFormatter("NodaTime.LocalDate", "yyyy-MM-dd");
+        /// <summary>
+        /// Formats LocalDate as yyyy-MM-dd.
+        /// </summary>
+        public static IValueFormatter LocalDateFormatter { get; } = new FormattableFormatter("NodaTime.LocalDate", "yyyy-MM-dd");
 
-        public static IValueFormatter LocalTimeFormatter =
-            new FormattableFormatter("NodaTime.LocalTime", "HH:mm:ss");
+        /// <summary>
+        /// Formats LocalTime as HH:mm:ss.
+        /// </summary>
+        public static IValueFormatter LocalTimeFormatter { get; } = new FormattableFormatter("NodaTime.LocalTime", "HH:mm:ss");
 
-        public static IValueFormatter LocalDateTimeFormatter =
-            new FormattableFormatter("NodaTime.LocalDateTime", "yyyy-MM-ddTHH:mm:ss");
+        /// <summary>
+        /// Formats LocalDateTime as yyyy-MM-ddTHH:mm:ss.
+        /// </summary>
+        public static IValueFormatter LocalDateTimeFormatter { get; } = new FormattableFormatter("NodaTime.LocalDateTime", "yyyy-MM-ddTHH:mm:ss");
 
-        public static IValueFormatter DurationFormatter =
-            new FormattableFormatter("NodaTime.Duration", "-H:mm:ss.FFFFFFFFF");
+        /// <summary>
+        /// Formats Duration as -H:mm:ss.FFFFFFFFF (without days).
+        /// </summary>
+        public static IValueFormatter DurationFormatter { get; } = new FormattableFormatter("NodaTime.Duration", "-H:mm:ss.FFFFFFFFF");
 
-        public static IValueFormatter DurationWithDaysFormatter =
-            new FormattableFormatter("NodaTime.Duration", "-D:hh:mm:ss.FFFFFFFFF");
+        /// <summary>
+        /// Formats Duration as -D:hh:mm:ss.FFFFFFFFF (with days).
+        /// </summary>
+        public static IValueFormatter DurationWithDaysFormatter { get; } = new FormattableFormatter("NodaTime.Duration", "-D:hh:mm:ss.FFFFFFFFF");
 
         /// <inheritdoc />
         public IEnumerable<IValueFormatter> GetFormatters()